Ported LETTER to Python
pretty straightforward port of the BASIC code to Python. Added a few functions for clarity. Removed CHR$(7) sounding of the bell, which is likely unsupported, and likely annoying.

Also corrected a mis-scanned "Let's Plan Again" message in the .BAS file, to both make more sense and also agree with the 2nd edition printed book.