Port AWARI to Python
I attempted to do a faithful port of the logic for the most part. In particular, there's a clever base-6 encoding of move history that allows games to be stored as integers in a "losing book" array, which allows the AI to become stronger over time. I did not change this logic, as I felt that the cleverness of this solution is worth observing. I endeavored to make every variable clearer, and use functions to help make program flow easier to understand.

I did make one small change to the logic, which I also documented clearly, having to do with evaluating very large moves. The code, as originally published, looks like a typo to me, but with BASIC, it's hard to be certain.

Also, I'm bad at this game; the computer beats me almost always.