update port CHECKERS to python
Moved a lot of methods into the board class, documented the logic for the board evaluation heuristics.

Documented some known issues with the original BASIC source code.

I've noticed that the computer has resigned to me when it has legal, if bad, moves. A bug in my code? A bug in the BASIC code? Still need to investigate further.