Python solution for 50: Horserace
Avoided any dictionary types, since those are not available in the
original BASIC version. This makes it slightly less idiomatic.
Tried my best to keep it as close to the original as possible.

Also horses are enumerated from 0, rather than from 1 as in the
original implementation (if needed this can easily be adjusted).