Revert "Fix bug in Lunar: wrong capsule weight."
This reverts commit 52e1bf49984960f0030f81bf968029c78b9729c9.

The original commit modified the game away from what was in BASIC Computer
Games, replacing it with the value from the original FOCAL version.  The
difference was intentional, as the introductory text makes clear, which is also
in the README.md here:

"To make the landing more of a challenge, but more closely approximate the real
Apollo LEM capsule, you should make the available fuel at the start (N) equal to
16,000 lbs, and the weight of the capsule (M) equal to 32,500 lbs."

Here we bring the value back into line with the BASIC Computer Games book & the
README.md text.  If we wanted to keep the FOCAL value, we should update the text
in the README.md to explain why.

See issue #867